// Backend/src/molecules/parser/parser.types.ts
export type atomsInfoIndex = {
    id_i: number;
    type_i: number;
    a_i: number;
    b_i: number;
    c_i: number;
    x_i: number;
    y_i: number;
    z_i: number;
    atoms_start_i: number;
}
export type bondsInfoIndex = {
    atom1_i: number;
    atom2_i: number;
    order_i: number;
    bonds_start_i: number;
}